restructure: create separate projects for core, database and api respectively
