+++ backend/Database/ValueConverters/DateOnlyConverter.cs
@@ -1,14 +1,14 @@
-using System.Linq.Expressions;
 using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 
-namespace backend.Database.ValueConverters;
+namespace Database.ValueConverters;
 
 public class DateOnlyConverter : ValueConverter<DateOnly, DateTime>
 {
-    public DateOnlyConverter() 
+    public DateOnlyConverter()
         : base(
             dateOnly => dateOnly.ToDateTime(TimeOnly.MinValue),
             dateTime => DateOnly.FromDateTime(dateTime)
-            )
-    { }
+        )
+    {
+    }
 }
