--- backend/WeatherForecast.cs
@@ -1,6 +1,0 @@
-namespace backend;
-
-public class WeatherForecast
-{
-    public int TemperatureC { get; set; }
-}
