--- backend/backend.csproj
@@ -1,60 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-    <PropertyGroup>
-        <TargetFramework>net7.0</TargetFramework>
-        <Nullable>enable</Nullable>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <UserSecretsId>aspnet-backend-79E3B0E8-84F2-45AB-94ED-4DA13F302B03</UserSecretsId>
-        <DockerDefaultTargetOS>Linux</DockerDefaultTargetOS>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="7.0.10" />
-        <PackageReference Include="Microsoft.AspNetCore.Authentication.OpenIdConnect" Version="7.0.10" />
-        <PackageReference Include="Microsoft.AspNetCore.OpenApi" Version="7.0.10" />
-        <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="7.0.10">
-            <PrivateAssets>all</PrivateAssets>
-            <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-        </PackageReference>
-        <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="7.0.10" />
-        <PackageReference Include="Microsoft.Identity.Web" Version="1.16.0" />
-        <PackageReference Include="Microsoft.OpenApi" Version="1.6.8" />
-        <PackageReference Include="PublicHoliday" Version="2.31.0" />
-        <PackageReference Include="Swashbuckle.AspNetCore" Version="6.5.0" />
-        <PackageReference Include="Swashbuckle.AspNetCore.Filters" Version="7.0.11" />
-    </ItemGroup>
-
-
-    <ItemGroup>
-        <Content Update="appsettings.Local.json.template">
-            <ExcludeFromSingleFile>true</ExcludeFromSingleFile>
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-            <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-        </Content>
-        <Content Remove="Tests\**" />
-    </ItemGroup>
-
-
-    <ItemGroup>
-        <Compile Remove="Tests\**" />
-    </ItemGroup>
-
-
-    <ItemGroup>
-        <EmbeddedResource Remove="Tests\**" />
-    </ItemGroup>
-
-
-    <ItemGroup>
-        <None Remove="Tests\**" />
-    </ItemGroup>
-
-
-    <ItemGroup>
-        <_ContentIncludedByDefault Remove="Tests\obj\project.assets.json" />
-        <_ContentIncludedByDefault Remove="Tests\obj\project.packagespec.json" />
-        <_ContentIncludedByDefault Remove="Tests\obj\Tests.csproj.nuget.dgspec.json" />
-    </ItemGroup>
-
-
-</Project>
