+++ backend/backend.sln
@@ -1,8 +1,12 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "backend", "backend.csproj", "{4E8F5F60-DBE5-4EF9-90C5-AB3C43E4ABC2}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tests", "Tests\Tests.csproj", "{05171493-9851-4302-B9C5-3053519A5A58}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Core", "Core\Core.csproj", "{208E55A4-D3F9-4425-A40F-F28570D75CA3}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Api", "Api\Api.csproj", "{3BBC1336-1810-4D80-8414-7F2AD78ABB23}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Database", "Database\Database.csproj", "{00163081-FB16-42B5-9C8F-2739CBC4176E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -10,13 +14,21 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{4E8F5F60-DBE5-4EF9-90C5-AB3C43E4ABC2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4E8F5F60-DBE5-4EF9-90C5-AB3C43E4ABC2}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4E8F5F60-DBE5-4EF9-90C5-AB3C43E4ABC2}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4E8F5F60-DBE5-4EF9-90C5-AB3C43E4ABC2}.Release|Any CPU.Build.0 = Release|Any CPU
 		{05171493-9851-4302-B9C5-3053519A5A58}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{05171493-9851-4302-B9C5-3053519A5A58}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{05171493-9851-4302-B9C5-3053519A5A58}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{05171493-9851-4302-B9C5-3053519A5A58}.Release|Any CPU.Build.0 = Release|Any CPU
+		{208E55A4-D3F9-4425-A40F-F28570D75CA3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{208E55A4-D3F9-4425-A40F-F28570D75CA3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{208E55A4-D3F9-4425-A40F-F28570D75CA3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{208E55A4-D3F9-4425-A40F-F28570D75CA3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3BBC1336-1810-4D80-8414-7F2AD78ABB23}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3BBC1336-1810-4D80-8414-7F2AD78ABB23}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3BBC1336-1810-4D80-8414-7F2AD78ABB23}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3BBC1336-1810-4D80-8414-7F2AD78ABB23}.Release|Any CPU.Build.0 = Release|Any CPU
+		{00163081-FB16-42B5-9C8F-2739CBC4176E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{00163081-FB16-42B5-9C8F-2739CBC4176E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{00163081-FB16-42B5-9C8F-2739CBC4176E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{00163081-FB16-42B5-9C8F-2739CBC4176E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
